Add create event route

+++ backend/db/migrations/20240402063923-create-event.js
@@ -39,7 +39,7 @@ module.exports = {
         allowNull: false
       },
       price: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DECIMAL(6, 2),
         allowNull: false
       },
       startDate: {
